requirements.txt
Django==3.1.4
django-cors-headers==3.5.0
djangorestframework==3.12.2
mysqlclient==2.0.1
django-mysql==3.9
PyJWT==1.7.1
django-cors-headers==3.5.0